Move database methods back into db.rb
