+++ lib/task-manager/employee.rb
@@ -7,15 +7,9 @@ class TM::Employee
     @name = name
     @employee_id = TM::Employee.gen_id
     @@employeecount += 1
-    @tasks = []
   end
 
   def self.gen_id
     @@employeecount
   end
-
-  def taketask(task)
-    @tasks.push(task)
-  end
-
 end
